api/requirements.txt
-i https://pypi.python.org/simple
aiohttp==3.8.3
aiosignal==1.2.0 ; python_version >= '3.6'
anyio==3.6.2 ; python_full_version >= '3.6.2'
asgiref==3.5.2 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==22.1.0 ; python_version >= '3.5'
backoff==2.2.1 ; python_version >= '3.7'
bcrypt==4.0.1
cassandra-driver==3.25.0
certifi==2022.9.24 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
click==8.1.3 ; python_version >= '3.7'
cryptography==38.0.1
deprecated==1.2.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ecdsa==0.18.0 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
elastic-transport==8.4.0 ; python_version >= '3.6'
elasticsearch==8.4.3
fastapi==0.85.1
frozenlist==1.3.1 ; python_version >= '3.7'
geomet==0.2.1.post1 ; python_version >= '2.7' and python_version != '3.3' and python_version < '4'
googleapis-common-protos==1.56.4 ; python_version >= '3.7'
grpcio==1.50.0 ; python_version >= '3.7'
h11==0.14.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
multidict==6.0.2 ; python_version >= '3.7'
neo4j-driver==4.3.6 ; python_version >= '3.5'
neobolt==1.7.17
neomodel==4.0.8
opentelemetry-api==1.13.0
opentelemetry-exporter-otlp==1.13.0
opentelemetry-exporter-otlp-proto-grpc==1.13.0 ; python_version >= '3.7'
opentelemetry-exporter-otlp-proto-http==1.13.0 ; python_version >= '3.7'
opentelemetry-instrumentation==0.34b0 ; python_version >= '3.7'
opentelemetry-instrumentation-asgi==0.34b0 ; python_version >= '3.7'
opentelemetry-instrumentation-fastapi==0.34b0
opentelemetry-instrumentation-logging==0.34b0
opentelemetry-proto==1.13.0 ; python_version >= '3.7'
opentelemetry-sdk==1.13.0
opentelemetry-semantic-conventions==0.34b0 ; python_version >= '3.7'
opentelemetry-util-http==0.34b0 ; python_version >= '3.7'
passlib[bcrypt]==1.7.4
prometheus-client==0.15.0
protobuf==3.20.3 ; python_version >= '3.7'
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pyhumps==3.7.3
python-jose[cryptography]==3.3.0
pytz==2022.5
requests==2.28.1 ; python_version >= '3.7' and python_version < '4'
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
sentry-sdk[fastapi]==1.10.1
setuptools==65.5.0 ; python_version >= '3.7'
shapely==1.7.1
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.20.4 ; python_version >= '3.7'
typing-extensions==4.4.0 ; python_version >= '3.7'
urllib3==1.26.12 ; python_version >= '3.6'
uvicorn==0.18.3
wrapt==1.14.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
yarl==1.8.1 ; python_version >= '3.7'
